pyproject.toml
[project]
name = "tw-stock-monitor"
version = "0.1.0"
description = "Intraday TW stock monitor that triggers Telegram alerts on technical + volume rules."
requires-python = ">=3.10"
dependencies = [
  "shioaji>=1.2",
  "python-telegram-bot>=21",
  "pyyaml>=6",
  "python-dotenv>=1",
  "loguru>=0.7",
  "pandas>=2.0",
  "numpy>=1.24",
]

[project.optional-dependencies]
dev = [
  "pytest>=8",
  "pytest-asyncio>=0.23",
  "ruff>=0.5",
]
ib = [
  # Interactive Brokers integration for overseas futures (CME/NYMEX/COMEX/…).
  # ib_async is the actively-maintained fork of ib_insync — it fixes the
  # Python 3.12+ event-loop incompatibility (eventkit calls
  # asyncio.get_event_loop() at import time).
  # Lazy-imported by monitor.broker.ib_client; install only if you watch
  # overseas_futures symbols.
  "ib_async>=1.0.0",
]

[project.scripts]
monitor = "monitor.app:main"
version = "monitor.testing_flow:show_version"
stock_testing = "monitor.testing_flow:testing_stock_ordering"
futures_testing = "monitor.testing_flow:testing_futures_ordering"
indicators_demo = "monitor.demo:indicators_demo"
monitor-backtest = "monitor.backtest.cli:backtest_cli"
overseas_quote_check = "monitor.broker.ib_quote_check:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"
